requirements.txt
aiohttp==3.9.5
aiofiles==24.1.0
colorlog==6.8.2
requests==2.32.3
fake-useragent==1.5.1
python-dotenv==1.0.1
